+++ apps/backend/src/modules/users/libs/validation-schemas/validation-schemas.ts
@@ -1,5 +1,7 @@
 export {
+	userForgotPasswordVaidationSchema,
 	userPreferencesValidationSchema,
+	userResetPasswordValidationSchema,
 	userSignInValidationSchema,
 	userSignUpValidationSchema,
 } from "shared";