+++ apps/backend/src/modules/users/user.service.ts
@@ -19,6 +19,20 @@ class UserService implements Service {
 		this.encrypt = encrypt;
 	}
 
+	public async changePassword(
+		id: number,
+		password: string,
+	): Promise<UserEntity> {
+		const { hash, salt } = await this.encrypt.encrypt(password);
+
+		const updates = {
+			passwordHash: hash,
+			passwordSalt: salt,
+		};
+
+		return await this.userRepository.updatePassword(id, updates);
+	}
+
 	public async create(payload: UserSignUpRequestDto): Promise<UserDto> {
 		const { hash, salt } = await this.encrypt.encrypt(payload.password);
 
@@ -41,6 +55,7 @@ class UserService implements Service {
 	public async find(id: number): Promise<null | UserEntity> {
 		return await this.userRepository.find(id);
 	}
+
 	public async findAll(): Promise<UserGetAllResponseDto> {
 		const items = await this.userRepository.findAll();