+++ apps/frontend/src/modules/auth/slices/actions.ts
@@ -1,8 +1,11 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-import { storage, StorageKey } from "~/libs/modules/storage/storage.js";
+import { NotificationMessage } from "~/libs/enums/enums.js";
+import { StorageKey } from "~/libs/modules/storage/storage.js";
 import { type AsyncThunkConfig } from "~/libs/types/types.js";
 import {
+	type EmailDto,
+	type ResetPasswordDto,
 	type UserDto,
 	type UserSignInRequestDto,
 	type UserSignUpRequestDto,
@@ -15,7 +18,7 @@ const signIn = createAsyncThunk<
 	UserSignInRequestDto,
 	AsyncThunkConfig
 >(`${sliceName}/sign-in`, async (signInPayload, { extra }) => {
-	const { authApi } = extra;
+	const { authApi, storage } = extra;
 
 	const { token, user } = await authApi.signIn(signInPayload);
 
@@ -29,21 +32,29 @@ const signUp = createAsyncThunk<
 	UserSignUpRequestDto,
 	AsyncThunkConfig
 >(`${sliceName}/sign-up`, async (registerPayload, { extra }) => {
-	const { authApi } = extra;
-
+	const { authApi, storage } = extra;
 	const { token, user } = await authApi.signUp(registerPayload);
-
 	void storage.set(StorageKey.TOKEN, token);
 
 	return user;
 });
+
+const logOut = createAsyncThunk<null, undefined, AsyncThunkConfig>(
+	`${sliceName}/log-out`,
+	async (_, { extra }) => {
+		const { storage } = extra;
+		await storage.drop(StorageKey.TOKEN);
+
+		return null;
+	},
+);
 
 const getAuthenticatedUser = createAsyncThunk<
 	null | UserDto,
 	undefined,
 	AsyncThunkConfig
 >(`${sliceName}/get-authenticated-user`, async (_, { extra }) => {
-	const { authApi } = extra;
+	const { authApi, storage } = extra;
 
 	const token = await storage.get(StorageKey.TOKEN);
 	const hasToken = Boolean(token);
@@ -55,4 +66,41 @@ const getAuthenticatedUser = createAsyncThunk<
 	return await authApi.getAuthenticatedUser();
 });
 
-export { getAuthenticatedUser, signIn, signUp };
+const requestResetPassword = createAsyncThunk<
+	boolean,
+	EmailDto,
+	AsyncThunkConfig
+>(`${sliceName}/send-reset-password-link`, async (emailPayload, { extra }) => {
+	const { authApi, notification } = extra;
+
+	const isSuccessful = await authApi.requestResetPassword(emailPayload);
+
+	if (isSuccessful) {
+		notification.success(NotificationMessage.LINK_SENT);
+	}
+
+	return isSuccessful;
+});
+
+const resetPassword = createAsyncThunk<
+	UserDto,
+	ResetPasswordDto,
+	AsyncThunkConfig
+>(`${sliceName}/reset-password`, async (emailPayload, { extra }) => {
+	const { authApi, storage } = extra;
+
+	const { token, user } = await authApi.resetPassword(emailPayload);
+
+	void storage.set(StorageKey.TOKEN, token);
+
+	return user;
+});
+
+export {
+	getAuthenticatedUser,
+	logOut,
+	requestResetPassword,
+	resetPassword,
+	signIn,
+	signUp,
+};