+++ packages/shared/src/index.ts
@@ -58,11 +58,16 @@ export {
 	quizUserAnswersValidationSchema,
 } from "./modules/quiz/quiz.js";
 export {
+	type EmailDto,
+	type ResetPasswordDto,
+	type ResetPasswordFormDto,
 	type UserDto,
+	userForgotPasswordVaidationSchema,
 	type UserGetAllResponseDto,
 	type UserPreferencesPayloadDto,
 	type UserPreferencesRequestDto,
 	userPreferencesValidationSchema,
+	userResetPasswordValidationSchema,
 	UsersApiPath,
 	type UserSignInRequestDto,
 	type UserSignInResponseDto,
@@ -72,4 +77,5 @@ export {
 	type UserSignUpResponseDto,
 	userSignUpValidationSchema,
 	UserValidationMessage,
+	UserValidationRule,
 } from "./modules/users/users.js";