+++ packages/shared/src/modules/users/libs/enums/enums.ts
@@ -1,4 +1,5 @@
 export { UserPreferencesValidationRule } from "./user-preferences-validation-rule.enum.js";
 export { UserValidationMessage } from "./user-validation-message.enum.js";
+export { UserValidationRegexRule } from "./user-validation-regex-rule.enum.js";
 export { UserValidationRule } from "./user-validation-rule.enum.js";
 export { UsersApiPath } from "./users-api-path.enum.js";