+++ packages/shared/src/modules/users/libs/types/email-dto.type.ts
@@ -1,0 +1,5 @@
+type EmailDto = {
+	email: string;
+};
+
+export { type EmailDto };