+++ packages/shared/src/modules/users/libs/validation-schemas/validation-schemas.ts
@@ -1,3 +1,5 @@
+export { userForgotPassword } from "./user-forgot-password.validation-schema.js";
 export { userPreferences } from "./user-preferences-validation-schema.js";
+export { userResetPassword } from "./user-reset-password.validation-schema.js";
 export { userSignIn } from "./user-sign-in.validation-schema.js";
 export { userSignUp } from "./user-sign-up.validation-schema.js";