+++ packages/shared/src/modules/users/users.ts
@@ -1,5 +1,12 @@
-export { UsersApiPath, UserValidationMessage } from "./libs/enums/enums.js";
 export {
+	UsersApiPath,
+	UserValidationMessage,
+	UserValidationRule,
+} from "./libs/enums/enums.js";
+export {
+	type EmailDto,
+	type ResetPasswordDto,
+	type ResetPasswordFormDto,
 	type UserDto,
 	type UserGetAllResponseDto,
 	type UserPreferencesPayloadDto,
@@ -10,6 +17,10 @@ export {
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "./libs/types/types.js";
-export { userPreferences as userPreferencesValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
-export { userSignUp as userSignUpValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
-export { userSignIn as userSignInValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
+export {
+	userForgotPassword as userForgotPasswordVaidationSchema,
+	userPreferences as userPreferencesValidationSchema,
+	userResetPassword as userResetPasswordValidationSchema,
+	userSignIn as userSignInValidationSchema,
+	userSignUp as userSignUpValidationSchema,
+} from "./libs/validation-schemas/validation-schemas.js";